Move some utilities to `solver_utils.py`
